+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * Update record metadata object when saving circulation rules. Refs UICIRC-858.
 * Add extra fields to reminder fees policy. Refs UICIRC-955.
 * Add New setting to enable hold requests to fail. Refs UICIRC-949.
+* Fix Circ rules editor crashes when certain special characters entered in filter box. Refs UICIRC-921.
 
 ## [8.0.1](https://github.com/folio-org/ui-circulation/tree/v8.0.1) (2023-03-07)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v8.0.0...v8.0.1)